requirements_test.txt
# Dépendances pour les tests
colorlog==6.7.0
coverage==7.3.2
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
freezegun==1.2.2 